+++ lib/capybara/active_admin.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'capybara/active_admin/version'
+require 'capybara/active_admin/util'
 require 'capybara/active_admin/selectors'
 require 'capybara/active_admin/finders'
 require 'capybara/active_admin/matchers'