+++ lib/capybara/active_admin/finders/attributes_table.rb
@@ -4,7 +4,13 @@ module Capybara
   module ActiveAdmin
     module Finders
       module AttributesTable
-        # todo
+        # @param model_name [Class<Object>, nil] model name or class.
+        # @param record_id [String, Numeric, nil]
+        # @yield within attributes table
+        def within_attributes_table_for(model_name, record_id = nil)
+          selector = attributes_table_selector(model_name, record_id)
+          within(selector) { yield }
+        end
       end
     end
   end