+++ lib/capybara/active_admin/matchers/form.rb
@@ -4,27 +4,23 @@ module Capybara
   module ActiveAdmin
     module Matchers
       module Form
-        def have_form_error(error, options = {})
+        def have_form_error(text, options = {})
           field = options.delete(:field)
-          opts = options.merge(text: error)
-          return have_selector("li #{inline_error_selector}", opts) if field.nil?
+          opts = Util.options_with_text(text, options)
+          li_selector = input_container_selector field, options.slice(:exact)
 
-          label = find(label_selector, text: field)
-          li_id = label.ancestor('li')[:id]
-          have_selector("li##{li_id} #{inline_error_selector}", opts)
+          have_selector("#{li_selector} #{inline_error_selector}", opts)
         end
 
         def have_no_form_errors(options = {})
           field = options.delete(:field)
-          return have_none_of_selectors(:css, "li #{inline_error_selector}", options) if field.nil?
+          li_selector = input_container_selector field, options.slice(:exact)
 
-          label = find(label_selector, text: field)
-          li_id = label.ancestor('li')[:id]
-          have_none_of_selectors(:css, "li##{li_id} #{inline_error_selector}", options)
+          have_none_of_selectors(:css, "#{li_selector} #{inline_error_selector}", options)
         end
 
-        def have_semantic_error(error, options = {})
-          opts = options.merge(text: error)
+        def have_semantic_error(text, options = {})
+          opts = Util.options_with_text(text, options)
           have_selector(semantic_error_selector, opts)
         end
 
@@ -32,8 +28,8 @@ module Capybara
           have_selector(semantic_error_selector, options)
         end
 
-        def have_no_input(label, options = {})
-          opts = options.merge(text: label)
+        def have_no_input(text, options = {})
+          opts = Util.options_with_text(text, options)
           have_none_of_selectors(:css, label_selector, opts)
         end
 
@@ -42,8 +38,8 @@ module Capybara
           have_selector(selector, options)
         end
 
-        def have_input(label, options = {})
-          selector = input_selector label, options.slice(:type, :text)
+        def have_input(text, options = {})
+          selector = input_selector text, options.slice(:type, :text)
           opts = options.except(:type, :text)
           have_selector(selector, opts)
         end