+++ lib/capybara/active_admin/selectors/form.rb
@@ -44,6 +44,19 @@ module Capybara
           selector = %(#{selector}[value="#{text}"]) unless text.nil?
           selector
         end
+
+        def input_container_selector(label, exact: nil)
+          return 'li' if label.nil?
+
+          label_opts = Util.options_with_text(label, exact: exact)
+          label_node = find(label_selector, label_opts)
+          li_id = label_node.ancestor('li')[:id]
+          "li##{li_id}"
+        end
+
+        def filter_form_selector
+          '.filter_form'
+        end
       end
     end
   end