+++ lib/capybara/active_admin/selectors/layout.rb
@@ -21,6 +21,50 @@ module Capybara
 
           '.flashes .flash'
         end
+
+        def footer_selector
+          'div.footer#footer'
+        end
+
+        def tab_header_link_selector
+          '.tabs.ui-tabs li.ui-tabs-tab a'
+        end
+
+        def tab_content_selector
+          '.tab-content'
+        end
+
+        def sidebar_selector
+          '#sidebar'
+        end
+
+        def panel_selector
+          '.panel'
+        end
+
+        def panel_title_selector
+          'h3'
+        end
+
+        def panel_content_selector
+          '.panel_contents'
+        end
+
+        def batch_actions_button_selector
+          'div.batch_actions_selector'
+        end
+
+        def dropdown_list_selector
+          'ul.dropdown_menu_list'
+        end
+
+        def batch_action_selector(title)
+          "li a[data-action='#{title}']"
+        end
+
+        def modal_dialog_selector
+          '.active_admin_dialog'
+        end
       end
     end
   end